+++ src/interface/MATLAB/CHANGES.md
@@ -3,7 +3,7 @@
 This project follows [Semantic Versioning](https://semver.org/). 
 To access the latest release of the package, visit [the ParaMonte GitHub repository release page](https://github.com/cdslaborg/paramonte/releases).  
 
-## **Version 1.x.x**  
+## **Version 2.x.x**  
 
 ### Version  2.0.1 -- September 26, 2020  
 
@@ -11,7 +11,10 @@ To access the latest release of the package, visit [the ParaMonte GitHub reposit
 
 +   The minor bug in GridPlot class method `rotateAxesLabels()` that caused 
     the `readSample()` , `readChain()`, `readMarkovChain()` to crash upon adding 
-    Grid plots is now fixed.
+    Grid plots is now fixed.  
+
++   The minor bug in the naming of the ParaMonte kernel library files on macOS 
+    (Darwin) is now fixed.  
 
 ### Version  2.0.0 -- September 22, 2020